+++ AdventOfCode.2019/Common/IntCode/Models/IntCodeOutput.cs
@@ -4,8 +4,8 @@ namespace AdventOfCode._2019.Common.IntCode.Models
 {
     public class IntCodeOutput
     {
-        public IEnumerable<int> CurrentState { get; set; }
-        public int? Output { get; set; }
+        public IEnumerable<long> CurrentState { get; set; }
+        public long? Output { get; set; }
         public bool IsComplete { get; set; }
     }
 }
