+++ AdventOfCode.2019/Common/IntCode/Parser.cs
@@ -5,9 +5,9 @@ namespace AdventOfCode._2019.Common.IntCode
 {
     public static class Parser
     {
-        public static List<int> ParseIntCode(string intCode)
+        public static List<long> ParseIntCode(string intCode)
         {
-            return intCode.Split(',').Select(int.Parse).ToList();
+            return intCode.Split(',').Select(long.Parse).ToList();
         }
     }
 }
