+++ AdventOfCode.2019/Day5/Day5.cs
@@ -8,10 +8,10 @@ namespace AdventOfCode._2019.Day5
 {
     public static class Day5
     {
-        public static int Part1()
+        public static long Part1()
         {
             var lines = FileReader.ReadInputLines(5).ToList();
-            var inputs = new Queue<int>().WithValues(1);
+            var inputs = new Queue<long>().WithValues(1);
             var output = IntCodeLogic.ParseAndRunIntCode(lines.First(), inputs).ToEnumerable().ToList();
             
             if (output.Count(x => !x.IsComplete && x.Output != 0) != 1)
@@ -22,10 +22,10 @@ namespace AdventOfCode._2019.Day5
             return output.Single(x => x.Output != 0 && !x.IsComplete).Output ?? -1;
         }
 
-        public static int Part2()
+        public static long Part2()
         {
             var lines = FileReader.ReadInputLines(5).ToList();
-            var inputs = new Queue<int>().WithValues(5);
+            var inputs = new Queue<long>().WithValues(5);
             var output = IntCodeLogic.ParseAndRunIntCode(lines.First(), inputs).ToEnumerable();
 
             return output.Single(x => !x.IsComplete).Output ?? -1;
